modified eslint rule + added md5 hash lib + preferences menu item links to profile
