+++ .eslintrc
@@ -185,6 +185,11 @@
       {
         "extensions": [".js"]
       }
+    ],
+    // Generates false positives when props are immediately passed to other function
+    // Example: in a mapStateToProps function where prop is passed to selector function
+    "react/no-unused-prop-types": [
+      1 // warn
     ]
   }
 }
