+++ package.json
@@ -43,6 +43,7 @@
   "dependencies": {
     "i18next": "^10.4.1",
     "lodash": "^4.17.5",
+    "md5": "^2.2.1",
     "moment": "^2.21.0",
     "prop-types": "^15.6.1",
     "react": "^16.2.0",
